pyproject.toml
[project]
name = "money"
version = "0.1.0"
description = "Contracts & payments accounting MVP"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "SQLAlchemy>=2.0.30",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "email-validator>=2.1.0",
  "python-multipart>=0.0.9",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
]

[tool.ruff]
line-length = 100

